+++ core/src/dird/ua_status.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2001-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -266,12 +266,10 @@ static void DoAllStatus(UaContext* ua)
 
   DoDirectorStatus(ua);
 
+
   /* Count Storage items */
-  LockRes(my_config);
   i = 0;
-  foreach_res (store, R_STORAGE) {
-    i++;
-  }
+  foreach_res (store, R_STORAGE) { i++; }
   unique_store = (StorageResource**)malloc(i * sizeof(StorageResource));
   /* Find Unique Storage address/port */
   i = 0;
@@ -290,7 +288,6 @@ static void DoAllStatus(UaContext* ua)
       Dmsg2(40, "Stuffing: %s:%d\n", store->address, store->SDport);
     }
   }
-  UnlockRes(my_config);
 
   previous_JobStatus = ua->jcr->getJobStatus();
 
@@ -301,12 +298,11 @@ static void DoAllStatus(UaContext* ua)
   }
   free(unique_store);
 
+
   /* Count Client items */
-  LockRes(my_config);
+
   i = 0;
-  foreach_res (client, R_CLIENT) {
-    i++;
-  }
+  foreach_res (client, R_CLIENT) { i++; }
   unique_client = (ClientResource**)malloc(i * sizeof(ClientResource));
   /* Find Unique Client address/port */
   i = 0;
@@ -325,7 +321,6 @@ static void DoAllStatus(UaContext* ua)
       Dmsg2(40, "Stuffing: %s:%d\n", client->address, client->FDport);
     }
   }
-  UnlockRes(my_config);
 
   previous_JobStatus = ua->jcr->getJobStatus();
 
@@ -394,30 +389,24 @@ static bool show_scheduled_preview(UaContext*,
       tm.tm_min = run->minute;         /* Set run minute */
       tm.tm_sec = 0;                   /* Zero secs */
 
-      /*
-       * Convert the time into a user parsable string.
+      /* Convert the time into a user parsable string.
        * As we use locale specific strings for weekday and month we
-       * need to keep track of the longest data string used.
-       */
+       * need to keep track of the longest data string used. */
       runtime = mktime(&tm);
       bstrftime_wd(dt, sizeof(dt), runtime);
       date_len = strlen(dt);
       if (date_len > *max_date_len) {
         if (*max_date_len == 0) {
-          /*
-           * When the datelen changes during the loop the locale generates a
+          /* When the datelen changes during the loop the locale generates a
            * date string that is variable. Only thing we can do about that is
            * start from scratch again. We invoke this by return false from this
-           * function.
-           */
+           * function. */
           *max_date_len = date_len;
           PmStrcpy(overview, "");
           return false;
         } else {
-          /*
-           * This is the first determined length we use this until we are proven
-           * wrong.
-           */
+          /* This is the first determined length we use this until we are proven
+           * wrong. */
           *max_date_len = date_len;
         }
       }
@@ -471,10 +460,8 @@ static bool show_scheduled_preview(UaContext*,
     }
   }
 
-  /*
-   * If we make it till here the length of the datefield is constant or didn't
-   * change.
-   */
+  /* If we make it till here the length of the datefield is constant or didn't
+   * change. */
   return true;
 }
 
@@ -658,7 +645,7 @@ static void DoSchedulerStatus(UaContext* ua)
   ua->SendMsg("Schedule               Jobs Triggered\n");
   ua->SendMsg("===========================================================\n");
 
-  LockRes(my_config);
+
   foreach_res (sched, R_SCHEDULE) {
     int cnt = 0;
 
@@ -702,7 +689,7 @@ static void DoSchedulerStatus(UaContext* ua)
 
     if (cnt > 0) { ua->SendMsg("\n"); }
   }
-  UnlockRes(my_config);
+
 
   // Build an overview.
   if (days > 0) { /* future */
@@ -729,7 +716,6 @@ start_again:
           }
         }
       } else {
-        LockRes(my_config);
         foreach_res (job, R_JOB) {
           if (!ua->AclAccessOk(Job_ACL, job->resource_name_)) { continue; }
 
@@ -741,17 +727,16 @@ start_again:
             if (!show_scheduled_preview(ua, job->schedule, overview,
                                         &max_date_len, time_to_check)) {
               job = NULL;
-              UnlockRes(my_config);
               goto start_again;
             }
           }
         }
-        UnlockRes(my_config);
+
         job = NULL;
       }
     } else {
       // List all schedules.
-      LockRes(my_config);
+
       foreach_res (sched, R_SCHEDULE) {
         if (!sched->enabled) { continue; }
 
@@ -763,11 +748,9 @@ start_again:
 
         if (!show_scheduled_preview(ua, sched, overview, &max_date_len,
                                     time_to_check)) {
-          UnlockRes(my_config);
           goto start_again;
         }
       }
-      UnlockRes(my_config);
     }
 
     time_to_check += seconds_per_hour; /* next hour */
@@ -913,7 +896,6 @@ static void ListScheduledJobs(UaContext* ua)
   }
 
   // Loop through all jobs
-  LockRes(my_config);
   foreach_res (job, R_JOB) {
     if (!ua->AclAccessOk(Job_ACL, job->resource_name_) || !job->enabled
         || (job->client && !job->client->enabled)) {
@@ -947,10 +929,8 @@ static void ListScheduledJobs(UaContext* ua)
       num_jobs++;
     }
   } /* end for loop over resources */
-  UnlockRes(my_config);
-  foreach_dlist (sp, sched) {
-    PrtRuntime(ua, sp);
-  }
+
+  foreach_dlist (sp, sched) { PrtRuntime(ua, sp); }
   if (num_jobs == 0 && !ua->api) { ua->SendMsg(_("No Scheduled Jobs.\n")); }
   if (!ua->api) ua->SendMsg("====\n");
   Dmsg0(200, "Leave list_sched_jobs_runs()\n");
@@ -1457,10 +1437,8 @@ static void StatusContentApi(UaContext* ua, StorageResource* store)
             }
             break;
           case slot_status_t::kSlotStatusEmpty:
-            /*
-             * See if this empty slot is empty because the volume is loaded
-             * in one of the drives.
-             */
+            /* See if this empty slot is empty because the volume is loaded
+             * in one of the drives. */
             vl2 = vol_is_loaded_in_drive(store, vol_list,
                                          vl1->bareos_slot_number);
             if (vl2) {
@@ -1562,10 +1540,8 @@ static void StatusContentJson(UaContext* ua, StorageResource* store)
             }
             break;
           case slot_status_t::kSlotStatusEmpty:
-            /*
-             * See if this empty slot is empty because the volume is loaded
-             * in one of the drives.
-             */
+            /* See if this empty slot is empty because the volume is loaded
+             * in one of the drives. */
             vl2 = vol_is_loaded_in_drive(store, vol_list,
                                          vl1->bareos_slot_number);
             if (vl2) {
@@ -1664,10 +1640,8 @@ static void StatusSlots(UaContext* ua, StorageResource* store)
       _("------+------------------+-----------+----------------+---------------"
         "-----------|\n"));
 
-  /*
-   * Walk through the list getting the media records
-   * Slots start numbering at 1.
-   */
+  /* Walk through the list getting the media records
+   * Slots start numbering at 1. */
   foreach_dlist (vl1, vol_list->contents) {
     vl2 = NULL;
     switch (vl1->slot_type) {
@@ -1690,10 +1664,8 @@ static void StatusSlots(UaContext* ua, StorageResource* store)
         switch (vl1->slot_status) {
           case slot_status_t::kSlotStatusEmpty:
             if (vl1->slot_type == slot_type_t::kSlotTypeStorage) {
-              /*
-               * See if this empty slot is empty because the volume is loaded
-               * in one of the drives.
-               */
+              /* See if this empty slot is empty because the volume is loaded
+               * in one of the drives. */
               vl2 = vol_is_loaded_in_drive(store, vol_list,
                                            vl1->bareos_slot_number);
               if (!vl2) {
@@ -1708,10 +1680,8 @@ static void StatusSlots(UaContext* ua, StorageResource* store)
             }
             FALLTHROUGH_INTENDED;
           case slot_status_t::kSlotStatusFull: {
-            /*
-             * We get here for all slots with content and for empty
-             * slots with their volume loaded in a drive.
-             */
+            /* We get here for all slots with content and for empty
+             * slots with their volume loaded in a drive. */
             MediaDbRecord mr;
             if (vl1->slot_status == slot_status_t::kSlotStatusFull) {
               if (!vl1->VolName) {