+++ core/src/filed/filed.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2010 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -307,7 +307,7 @@ static bool CheckResources()
   DirectorResource* director;
   const std::string& configfile = my_config->get_base_config_path();
 
-  LockRes(my_config);
+  ResLocker _{my_config};
 
   me = (ClientResource*)my_config->GetNextRes(R_CLIENT, nullptr);
   my_config->own_resource_ = me;
@@ -421,8 +421,7 @@ static bool CheckResources()
         }
       }
 
-      /*
-       * Crypto recipients. We're always included as a recipient.
+      /* Crypto recipients. We're always included as a recipient.
        * The symmetric session key will be encrypted for each of these readers.
        */
       me->pki_recipients = new alist<X509_KEYPAIR*>(10, not_owned_by_alist);
@@ -458,16 +457,13 @@ static bool CheckResources()
 
 
   /* Verify that a director record exists */
-  LockRes(my_config);
   director = (DirectorResource*)my_config->GetNextRes(R_DIRECTOR, nullptr);
-  UnlockRes(my_config);
+
   if (!director) {
     Emsg1(M_FATAL, 0, _("No Director resource defined in %s\n"),
           configfile.c_str());
     OK = false;
   }
-
-  UnlockRes(my_config);
 
   if (OK) {
     CloseMsg(nullptr);              /* close temp message handler */