+++ core/src/stored/butil.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2010 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -140,10 +140,8 @@ static bool setup_to_access_device(DeviceControlRecord* dcr,
 
   InitReservationsLock();
 
-  /*
-   * If no volume name already given and no bsr, and it is a file,
-   * try getting name from Filename
-   */
+  /* If no volume name already given and no bsr, and it is a file,
+   * try getting name from Filename */
   if (!VolumeName.empty()) {
     bstrncpy(VolName, VolumeName.c_str(), sizeof(VolName));
     if (VolumeName.size() >= MAX_NAME_LENGTH) {
@@ -264,7 +262,6 @@ static DeviceResource* find_device_res(char* archive_device_string,
   DeviceResource* device_resource;
 
   Dmsg0(900, "Enter find_device_res\n");
-  LockRes(my_config);
   foreach_res (device_resource, R_DEVICE) {
     Dmsg2(900, "Compare %s and %s\n", device_resource->archive_device_string,
           archive_device_string);
@@ -292,7 +289,6 @@ static DeviceResource* find_device_res(char* archive_device_string,
       }
     }
   }
-  UnlockRes(my_config);
 
   if (!found) {
     Pmsg2(0, _("Could not find device \"%s\" in config file %s.\n"),